server feat: add admin page

+++ server/controllers/userControl.js
@@ -1,6 +1,7 @@
 const mongoose = require("mongoose");
 const models = require("../models/model");
 const User = models.User;
+const Project = models.Project;
 
 const listUsers = async (req, res) => {
   try {
@@ -8,6 +9,8 @@ const listUsers = async (req, res) => {
     const usersList = users.map((user) => ({
       id: user.id,
       name: user.name,
+      email: user.email,
+      createdAt: user.createdAt,
     }));
     res.json(usersList);
   } catch (err) {
@@ -61,11 +64,23 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   try {
-    const removedUser = await User.findByIdAndRemove(req.params.id);
-
-    if (!removedUser) {
+    const user = await User.findById(req.params.id);
+    if (!user) {
       return res.status(404).json({ message: "User not found" });
     }
+    if (!req.user.admin) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
+    const projects = await Project.deleteMany({ owner: req.params.id });
+    const userId = req.params.id;
+    await Project.updateMany(
+      {
+        $or: [{ members: userId }, { managers: userId }],
+      },
+      { $pull: { members: userId, managers: userId } }
+    );
+
+    const removedUser = await User.findByIdAndRemove(req.params.id);
 
     res.json(removedUser);
   } catch (err) {
